Update to use 1.8 EnumParticle class
Note that this IS NOT backwards compatible with 1.8!

I made a 1.7 branch for backwards compatibility, but moving forward
EffectLib will assume 1.8 support.

Note that subtype effects (blocks breaking, items cracking, etc) are
broken for now. I'm not certain how these are meant to work with the
new enum-based system yet.